Restore the RenderCallbackAdapter
It might still be useful to have this default implementation of a
RenderCallback.

It does use an intermediate int[] buffer for the video frame data,
and this is likely not the most efficient approach.

For direct access to the native video memory, a RenderCallback
should be used rather than this default implementation.